requirements.txt
flask>=2.0.0
sqlalchemy>=1.4.0
werkzeug>=2.0.0
jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
pandas>=1.3.0
xlsxwriter>=3.0.0
python-dotenv>=0.19.0
pytest>=6.2.0
flask-wtf>=1.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
cryptography>=35.0.0
faker>=8.0.0
python-dateutil>=2.8.0
pytz>=2021.1
pymongo>=3.12.0
redis>=4.0.0
bcrypt>=3.2.0
flask-sqlalchemy>=2.5.0
flask-login>=0.5.0
flask-migrate>=3.0.0
email-validator>=1.1.0
requests>=2.26.0